requirements.txt
streamlit>=1.2.0
tensorflow
Pillow
numpy
keras
scikit-learn
